requirements.txt
duckdb
pandas
psycopg2-binary
streamlit
plotly
python-dotenv